+++ js/test/name/testnamefmt.js
@@ -1,7 +1,7 @@
 /*
  * testnamefmt.js - test the name formatter object
  * 
- * Copyright © 2013-2015,2017, JEDLSoft
+ * Copyright © 2013-2015,2017,2019 JEDLSoft
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -165,7 +165,24 @@ module.exports.testnamefmt = {
         test.equal(fmt.format(name), "Mr. John Kevin Smith Phd.");
         test.done();
     },
-    
+
+    testNameFmtENFamiliar: function(test) {
+        test.expect(1);
+        var name = new Name({
+            prefix: "Mr.",
+            givenName: "John",
+            middleName: "Kevin",
+            familyName: "Smith",
+            suffix: "Phd."
+        });
+        var fmt = new NameFmt({
+            style: "familiar"
+        });
+
+        test.equal(fmt.format(name), "John");
+        test.done();
+    },
+
     testNameFmtENWithCommaInSuffix: function(test) {
         test.expect(1);
         var name = new Name({
@@ -383,7 +400,27 @@ module.exports.testnamefmt = {
         test.equal(fmt.format(name), "Hr. Andreas Helmut Schmidt MdB");
         test.done();
     },
-    
+
+    testNameFmtDEFamiliar: function(test) {
+        test.expect(1);
+        var name = new Name({
+            prefix: "Hr.",
+            givenName: "Andreas",
+            middleName: "Helmut",
+            familyName: "Schmidt",
+            suffix: "MdB"
+        }, {
+            locale: "de-DE"
+        });
+        var fmt = new NameFmt({
+            style: "familiar",
+            locale: "de-DE"
+        });
+
+        test.equal(fmt.format(name), "Hr. Schmidt");
+        test.done();
+    },
+
     testNameFmtDEWithCommaInSuffix: function(test) {
         test.expect(1);
         var name = new Name({
@@ -771,7 +808,7 @@ module.exports.testnamefmt = {
         test.done();
     },
 
-    testNameFmtZHFormalLong: function(test) {
+    testNameFmtKOFamiliar: function(test) {
         test.expect(1);
         var name = new Name({
             honorific: "닥터",
@@ -807,7 +844,78 @@ module.exports.testnamefmt = {
         test.equal(fmt.format(name), "닥터 박은성 님");
         test.done();
     },
-    
+
+    testNameFmtKOFormalLong: function(test) {
+        test.expect(1);
+        var name = new Name({
+            honorific: "닥터",
+            givenName: "은성",
+            familyName: "박",
+            suffix: "님"
+        }, {
+            locale: "ko-KR"
+        });
+        var fmt = new NameFmt({
+            style: "familiar",
+            locale: "ko-KR"
+        });
+
+        test.equal(fmt.format(name), "박은성 님");
+        test.done();
+    },
+    /*testNameFmtKOFamiliar2: function(test) {
+        test.expect(1);
+        var name = new Name({
+            givenName: "은성",
+            familyName: "박"
+        }, {
+            locale: "ko-KR"
+        });
+        var fmt = new NameFmt({
+            style: "familiar",
+            locale: "ko-KR"
+        });
+
+        test.equal(fmt.format(name), "박은성 님");
+        test.done();
+    },*/
+
+    testNameFmtCSFamiliar: function(test) {
+        test.expect(1);
+        var name = new Name({
+            honorific: "Pan",
+            givenName: "Jan",
+            familyName: "Novák"
+        }, {
+            locale: "cs-CZ"
+        });
+        var fmt = new NameFmt({
+            style: "familiar",
+            locale: "cs-CZ"
+        });
+
+        test.equal(fmt.format(name), "Pan Novák");
+        test.done();
+    },
+
+    testNameFmtSKFamiliar: function(test) {
+        test.expect(1);
+        var name = new Name({
+            honorific: "Pani",
+            givenName: "Mária",
+            familyName: "Obecny"
+        }, {
+            locale: "sk-SK"
+        });
+        var fmt = new NameFmt({
+            style: "familiar",
+            locale: "sk-SK"
+        });
+
+        test.equal(fmt.format(name), "Pani Obecny");
+        test.done();
+    },
+
     testNameFmtENWithImplicitConversionOfArgToName: function(test) {
         test.expect(1);
         var fmt = new NameFmt({